# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MATERION - Sistema de Compras ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /setores:
    get:
      operationId: listSetores
      tags: [setores]
      responses:
        "200":
          description: Lista de setores
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Setor" }
    post:
      operationId: createSetor
      tags: [setores]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetorInput" }
      responses:
        "201":
          description: Setor criado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Setor" }
  /setores/{id}:
    patch:
      operationId: updateSetor
      tags: [setores]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetorUpdate" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Setor" }
          description: Setor atualizado
    delete:
      operationId: deleteSetor
      tags: [setores]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Setor excluído

  # ─── CATEGORIAS ──────────────────────────────────────────────────────────────
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      parameters:
        - name: tipo
          in: query
          schema: { type: string }
        - name: fornecedorId
          in: query
          schema: { type: integer }
        - name: setorId
          in: query
          schema: { type: integer }
        - name: categoriaId
          in: query
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
        - name: ativo
          in: query
          schema: { type: boolean }
        - name: favorito
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Lista de templates
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OccTemplate" }
    post:
      operationId: createTemplate
      tags: [templates]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OccTemplateInput" }
      responses:
        "201":
          description: Template criado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OccTemplate" }
...
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OccTemplateComItens" }
          description: Template encontrado
        "404":
          description: Não encontrado
    patch:
      operationId: updateTemplate
      tags: [templates]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OccTemplateUpdate" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OccTemplate" }
          description: Template atualizado
    delete:
      operationId: deleteTemplate
      tags: [templates]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Template excluído
  /templates/{id}/favoritar:
    patch:
      operationId: favoritarTemplate
      tags: [templates]
      summary: Favoritar ou desfavoritar template
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FavoritarInput" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OccTemplate" }
          description: Template atualizado

  # ─── TEMPLATE ITENS ──────────────────────────────────────────────────────────
  /templates/{templateId}/itens:
    get:
      operationId: listTemplateItens
      tags: [templates]
      parameters:
        - { name: templateId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Itens do template
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TemplateItem" }
    post:
      operationId: createTemplateItem
      tags: [templates]
      parameters:
        - { name: templateId, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TemplateItemInput" }
      responses:
        "201":
          description: Item criado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TemplateItem" }
  /templates/{templateId}/itens/{id}:
    patch:
      operationId: updateTemplateItem
      tags: [templates]
      parameters:
        - { name: templateId, in: path, required: true, schema: { type: integer } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TemplateItemUpdate" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TemplateItem" }
          description: Item atualizado
    delete:
      operationId: deleteTemplateItem
      tags: [templates]
      parameters:
        - { name: templateId, in: path, required: true, schema: { type: integer } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Item excluído

  # ─── DASHBOARD ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Setor:
      type: object
      required: [id, nome, ativo, criadoEm]
      properties:
        id: { type: integer }
        nome: { type: string }
        ativo: { type: boolean }
        criadoEm: { type: string, format: date-time }
    SetorInput:
      type: object
      required: [nome]
      properties:
        nome: { type: string, minLength: 1 }
        ativo: { type: boolean }
    SetorUpdate:
      type: object
      properties:
        nome: { type: string }
        ativo: { type: boolean }

    # ─── Categoria ─────────────────────────────────────────────────────────────
...
    OccTemplate:
      type: object
      required: [id, nome, tipo, ativo, favorito, criadoEm]
      properties:
        id: { type: integer }
        nome: { type: string }
        descricao: { type: ["string", "null"] }
        tipo: { type: string }
        fornecedorPadraoId: { type: ["integer", "null"] }
        fornecedorPadraoNome: { type: ["string", "null"] }
        empresaId: { type: ["integer", "null"] }
        empresaNome: { type: ["string", "null"] }
        setorPadraoId: { type: ["integer", "null"] }
        setorPadraoNome: { type: ["string", "null"] }
        categoriaPadraoId: { type: ["integer", "null"] }
        categoriaPadraoNome: { type: ["string", "null"] }
        prioridade: { type: ["string", "null"] }
        observacoesPadrao: { type: ["string", "null"] }
        condicoesPagamento: { type: ["string", "null"] }
        prazoPadrao: { type: ["string", "null"] }
        layoutPdf: { type: ["string", "null"] }
        assinaturaRodape: { type: ["string", "null"] }
        ativo: { type: boolean }
        favorito: { type: boolean }
        usoCount: { type: integer }
        ultimoUso: { type: ["string", "null"], format: date-time }
        criadoEm: { type: string, format: date-time }
    OccTemplateComItens:
      type: object
      required: [id, nome, tipo, ativo, favorito, criadoEm, itens]
      properties:
        id: { type: integer }
        nome: { type: string }
        descricao: { type: ["string", "null"] }
        tipo: { type: string }
        fornecedorPadraoId: { type: ["integer", "null"] }
        fornecedorPadraoNome: { type: ["string", "null"] }
        empresaId: { type: ["integer", "null"] }
        empresaNome: { type: ["string", "null"] }
        setorPadraoId: { type: ["integer", "null"] }
        setorPadraoNome: { type: ["string", "null"] }
        categoriaPadraoId: { type: ["integer", "null"] }
        categoriaPadraoNome: { type: ["string", "null"] }
        prioridade: { type: ["string", "null"] }
        observacoesPadrao: { type: ["string", "null"] }
        condicoesPagamento: { type: ["string", "null"] }
        prazoPadrao: { type: ["string", "null"] }
        layoutPdf: { type: ["string", "null"] }
        assinaturaRodape: { type: ["string", "null"] }
        ativo: { type: boolean }
        favorito: { type: boolean }
        usoCount: { type: integer }
        ultimoUso: { type: ["string", "null"], format: date-time }
        criadoEm: { type: string, format: date-time }
        itens:
          type: array
          items: { $ref: "#/components/schemas/TemplateItem" }
    OccTemplateInput:
      type: object
      required: [nome, tipo]
      properties:
        nome: { type: string, minLength: 1 }
        descricao: { type: string }
        tipo: { type: string }
        fornecedorPadraoId: { type: integer }
        empresaId: { type: integer }
        setorPadraoId: { type: integer }
        categoriaPadraoId: { type: integer }
        prioridade: { type: string }
        observacoesPadrao: { type: string }
        condicoesPagamento: { type: string }
        prazoPadrao: { type: string }
        layoutPdf: { type: string }
        assinaturaRodape: { type: string }
        ativo: { type: boolean }
    OccTemplateUpdate:
      type: object
      properties:
        nome: { type: string }
        descricao: { type: string }
        tipo: { type: string }
        fornecedorPadraoId: { type: integer }
        empresaId: { type: integer }
        setorPadraoId: { type: integer }
        categoriaPadraoId: { type: integer }
        prioridade: { type: string }
        observacoesPadrao: { type: string }
        condicoesPagamento: { type: string }
        prazoPadrao: { type: string }
        layoutPdf: { type: string }
        assinaturaRodape: { type: string }
        ativo: { type: boolean }
    FavoritarInput:
      type: object
      required: [favorito]
      properties:
        favorito: { type: boolean }

    # ─── Template Item ─────────────────────────────────────────────────────────
    TemplateItem:
      type: object
      required: [id, templateId, nomeMaterial, unidade, quantidadePadrao]
      properties:
        id: { type: integer }
        templateId: { type: integer }
        materialId: { type: ["integer", "null"] }
        nomeMaterial: { type: string }
        unidade: { type: string }
        quantidadePadrao: { type: number }
        observacao: { type: ["string", "null"] }
        fornecedorPreferencialId: { type: ["integer", "null"] }
        fornecedorPreferencialNome: { type: ["string", "null"] }
    TemplateItemInput:
      type: object
      required: [nomeMaterial, unidade, quantidadePadrao]
      properties:
        materialId: { type: integer }
        nomeMaterial: { type: string, minLength: 1 }
        unidade: { type: string }
        quantidadePadrao: { type: number }
        observacao: { type: string }
        fornecedorPreferencialId: { type: integer }
    TemplateItemUpdate:
      type: object
      properties:
        materialId: { type: integer }
        nomeMaterial: { type: string }
        unidade: { type: string }
        quantidadePadrao: { type: number }
        observacao: { type: string }
        fornecedorPreferencialId: { type: integer }

    # ─── Dashboard ─────────────────────────────────────────────────────────────
...