# lib/api-zod/src/generated/types/favoritarInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MATERION - Sistema de Compras ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface FavoritarInput {
  favorito: boolean;
}