# lib/api-zod/src/generated/types/occTemplateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MATERION - Sistema de Compras ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface OccTemplateInput {
  /** @minLength 1 */
  nome: string;
  descricao?: string;
  tipo: string;
  fornecedorPadraoId?: number;
  empresaId?: number;
  setorPadraoId?: number;
  categoriaPadraoId?: number;
  prioridade?: string;
  observacoesPadrao?: string;
  condicoesPagamento?: string;
  prazoPadrao?: string;
  layoutPdf?: string;
  assinaturaRodape?: string;
  ativo?: boolean;
}