# lib/api-zod/src/generated/types/templateItemUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MATERION - Sistema de Compras ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface TemplateItemUpdate {
  materialId?: number;
  nomeMaterial?: string;
  unidade?: string;
  quantidadePadrao?: number;
  observacao?: string;
  fornecedorPreferencialId?: number;
}